pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scenecraft-engine"
version = "0.1.0"
description = "AI video editor"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
    "click>=8.0.0",
    "moviepy>=2.0.0",
    "Pillow>=10.0.0",
    "ruamel.yaml>=0.18.0",
]

[project.optional-dependencies]
ai = [
    "anthropic>=0.30.0",
]
google = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
]
kling = [
    "Pillow>=10.0.0",
]
describe = [
    "google-genai>=1.0.0",
]
describe-local = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.27.0",
]
all = [
    "scenecraft-engine[ai,describe]",
]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
scenecraft = "scenecraft.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
